+++ src/rpcServer/index.ts
@@ -1,1 +1,3 @@
 export * from "./RpcServer";
+export * from "./getCommunicationDirPath";
+export * from "./initializeCommunicationDir";
