+++ src/uninstall.ts
@@ -1,8 +1,8 @@
-import { getCommunicationDirPath } from "./paths";
 import { sync as rimrafSync } from "rimraf";
+import { getCommunicationDirPath } from "./rpcServer";
 
 function main() {
-  rimrafSync(getCommunicationDirPath(), { disableGlob: true });
+    rimrafSync(getCommunicationDirPath(), { disableGlob: true });
 }
 
 main();
